PLANET-5609 Move slideCovers function to CoversFrontend
That way we don't need to pass down the container ref to the useCovers hook, and it also makes more sense since it's the only place where it's used
