+++ assets/src/blocks/Covers/CoversFrontend.js
@@ -1,6 +1,6 @@
 import { Covers } from './Covers';
 import { useRef } from '@wordpress/element';
-import { COVERS_LAYOUTS } from './CoversConstants';
+import { COVERS_LAYOUTS, COVERS_TYPES } from './CoversConstants';
 import { useCovers } from './useCovers';
 import { CoversCarouselControls } from './CoversCarouselControls';
 import { CoversGridLoadMoreButton } from './CoversGridLoadMoreButton';
@@ -13,8 +13,8 @@ export const CoversFrontend = attributes => {
     showMoreCovers,
     row,
     amountOfCoversPerRow,
-    slideCovers,
-  } = useCovers(attributes, true, coversContainerRef);
+    setRow,
+  } = useCovers(attributes, true);
 
   const isCarouselLayout = layout === COVERS_LAYOUTS.carousel;
 
@@ -31,6 +31,38 @@ export const CoversFrontend = attributes => {
   if (!covers.length) {
     return null;
   }
+
+  const slideCovers = (direction, rowNumber) => {
+    if (!coversContainerRef.current) {
+      return;
+    }
+
+    const covers = coversContainerRef.current.querySelector('.covers');
+
+    if (!covers) {
+      return;
+    }
+
+    const isRTL = document.querySelector('html').dir === 'rtl';
+    // We need to account for the spacing between covers,
+    // in all styles that spacing is 24px for screens >= 992px.
+    let scrollOffset = (isRTL ? -1 : 1) * (covers.offsetWidth + 24);
+
+    // For the Take Action covers because of the box-shadow we had to add
+    // extra padding to the sides (5px each) so we need to account for them.
+    if (cover_type === COVERS_TYPES.takeAction) {
+      scrollOffset += (isRTL ? 1 : -1) * 10;
+    }
+
+    if (direction) {
+      const initialScrollPosition = covers.scrollLeft;
+      covers.scrollLeft = initialScrollPosition + (direction === 'next' ? scrollOffset : -scrollOffset);
+      setRow(direction === 'next' ? row + 1 : row - 1);
+    } else if (rowNumber && rowNumber !== row) {
+      covers.scrollLeft = (rowNumber - 1) * scrollOffset;
+      setRow(rowNumber);
+    }
+  };
 
   const showLoadMoreButton = !isCarouselLayout && !!initialRowsLimit && covers.length > (amountOfCoversPerRow * row);
